+++ modules/transport-grpc/src/main/java/org/opensearch/transport/grpc/proto/request/search/query/QueryBuilderProtoConverterRegistryImpl.java
@@ -50,7 +50,10 @@ public class QueryBuilderProtoConverterRegistryImpl implements QueryBuilderProto
         delegate.registerConverter(new TermsQueryBuilderProtoConverter());
         delegate.registerConverter(new MatchPhraseQueryBuilderProtoConverter());
         delegate.registerConverter(new MultiMatchQueryBuilderProtoConverter());
+        delegate.registerConverter(new BoolQueryBuilderProtoConverter());
 
+        // Set the registry on all converters so they can access each other
+        delegate.setRegistryOnAllConverters(this);
         logger.info("Registered {} built-in query converters", delegate.size());
     }
 
@@ -72,5 +75,14 @@ public class QueryBuilderProtoConverterRegistryImpl implements QueryBuilderProto
      */
     public void registerConverter(QueryBuilderProtoConverter converter) {
         delegate.registerConverter(converter);
+    }
+
+    /**
+     * Updates the registry on all registered converters.
+     * This should be called after all external converters have been registered
+     * to ensure converters like BoolQueryBuilderProtoConverter can access the complete registry.
+     */
+    public void updateRegistryOnAllConverters() {
+        delegate.setRegistryOnAllConverters(this);
     }
 }